significant refactoring of the library
