+++ universality/__init__.py
@@ -1,0 +1,2 @@
+"""a module housing routines for nonparametric Equation of State (EoS) inference based on Gaussian Processes"""
+__author__ = "Reed Essick (reed.essick@gmail.com)"
