+++ universality/gaussianprocess/__init__.py
@@ -1,0 +1,3 @@
+"""a module housing our logic for Gaussian Process construction, validation, and manipulation, including handling/optimizing hyperparameters
+"""
+__author__ = "Reed Essick (reed.essick@gmail.com)"
