+++ universality/kde/__init__.py
@@ -1,0 +1,7 @@
+"""a module housing our internal KDE logic
+"""
+__author__ = "Reed Essick (reed.essick@gmail.com)"
+
+#-------------------------------------------------
+
+from .kde import *
