+++ universality/plot/__init__.py
@@ -1,0 +1,3 @@
+"""a module that houses our custom plotting routines
+"""
+__author__ = "Reed Essick (reed.essick@gmail.com)"
