+++ universality/priors/__init__.py
@@ -1,0 +1,7 @@
+"""a module that houses logic to deal with different single-event parameter priors and astrophysical rates inference
+"""
+__author__ = "Reed Essick (reed.essick@gmail.com)"
+
+#-------------------------------------------------
+
+from .priors import *
