+++ universality/priors/rates.py
