+++ universality/properties/__init__.py
@@ -1,0 +1,11 @@
+"""a module that houses routines to extract both microphysical and macrophysical properties from EoS
+"""
+__author__ = "Reed Essick (reed.essick@gmail.com)"
+
+#-------------------------------------------------
+
+from .branches import *
+from .crossing import *
+from .extrema import *
+from .features import *
+from .samples import *
