+++ universality/properties/utils.py
@@ -1,0 +1,5 @@
+__doc__ = "a module for general utility functions for processing large sets of EoS to extract properties"
+__author__ = "reed.essick@ligo.org"
+
+#-------------------------------------------------
+
