+++ universality/stats/__init__.py
@@ -1,0 +1,7 @@
+"""a module that houses basic statistics we use to quantify our results based on large monte carlo integrals
+"""
+__author__ = "Reed Essick (reed.essick@gmail.com)"
+
+#-------------------------------------------------
+
+from .stats import *
