+++ universality/tov/__init__.py
@@ -1,0 +1,7 @@
+"""a module that houses our routines to solve the TOV equations, etc
+"""
+__author__ = "Reed Essick (reed.essick@gmail.com)"
+
+#-------------------------------------------------
+
+from .crust import *
