+++ universality/utils/units.py
@@ -1,0 +1,9 @@
+__doc__ = "definitions of units used throughout the library. Most calculations assume CGS"
+__author__ = "reed.essick@ligo.org"
+
+#-------------------------------------------------
+
+G = 6.674e-8        # newton's constant in (g^-1 cm^3 s^-2)
+c = (299792458*100) # speed of light in (cm/s)
+c2 = c**2
+Msun = 1.989e33     # mass of the sun in (g)
